requirements.txt
discord.py == 2.3.2
fpdf2 == 2.7.9
pdf2image == 1.17.0
openai == 1.30.1
